bootstrap/cert-manager: add patch to inject CA cert into correct namespace
With the bootstrap/cert-manager/kustomization.yaml we have overriden the
namespace of all the resources to be `kube-system`, but the cert-manager
still has the annotation `cert-manager.io/inject-ca-from-secret` that
refers to the default `cert-manager` namespace. As a result, it is not
possible for the cainjector to inject the secret.
This patch updates the annotation to use `kube-system` instead.