dex: add Azure authentication support
This commit brings the possibility of authenticating through Azure SSO
which should help with authenticating users.

Unfortunately, the groups functionality does not work in current format.
It boils down to the fact that `go-oidc` does not work with AADv1 for
groups (as per the https://dexidp.io/docs/connectors/oidc/). We are not
able to switch to AADv2 endpoint
(`https://login.microsoftonline.com...``) because the Issuer field has
to match exactly and the app has been created with Azure Portal (AADv1).

I have also tried several workarounds, which all failed but I thought it
maybe worth recording here:

- using `providerDiscoveryOverrides` to fake out all discovered URLs
  except the issuer. This is not possible, because the OIDC connector in dex
  allows only to override the token URL and authorization URL.

- setting `insecureEnableGroups:true` was the closest thing to having
  groups working. It makes the list of group UUIDs being put into the
  identity token. This list makes the token so large that it goes over
  the limit of the HTTP Request header value size for nginx, which is a
  problem when the Client attempts to do a call to
  https://dexidp.local/userinfo and would likely be a problem on
  subsequent requests too. The problem is partially because
  these are UUIDs (36 characters each) rather than the group names.

- use claimMapping of `roles` to `groups` similar to what we used in
  Nautobot, but that did not work since the AADv1 endpoint does not return
  them and AADv2 cannot be used due to restriction on the Issuer.

One potential alternative solution that I have not tried is using
Microsoft connector (https://dexidp.io/docs/connectors/microsoft/),
however it also has a set of limitations that make it possibly
unacceptable:
- requires Organization Admin access approval to provide access to the
  app to read the full Directory
- may suffer from the same AADv1 vs AADv2 problem
- as per docs: "Users that reject dex’s access through Microsoft will
  also revoke all dex clients which authenticated them through
  Microsoft."