replace Nautobot job with a workflow
Instead of waiting for an explicit button press, we setup Nautobot to
send webhooks to Argo Events any time the Interface is changed. When
particular conditions are met (device is a server, has the DRAC IP and
the Interface is named 'iLO'), we trigger the workflow to synchronize
information to Ironic.