fix: avoid ArgoCD job conflicts by replacing keystone jobs
Set the appropriate annotations on the keystone jobs by overriding the
ones that OpenStack Helm uses which are post-install / post-upgrade
which wait until the deployments are healthy to run. The deployments
cannot get healthy because they depend on those jobs to complete. By
letting the regular hooks be used, the jobs get replaced instead of
conflicting.