feat: switch back to nautobot helm chart
The upstream helm chart has been fixed with the patches we need so
there's no reason to include the rendered helm chart any longer.
Unfortunately there's a breaking change here around how the
NAUTOBOT_SECRET_KEY is handled which can only be supplied via a helm
variable and not via a reference.