feat: adjust hooks on Ironic jobs to avoid conflicts
Jobs cannot be run as pre-install or pre-upgrade because ArgoCD tries to
run them before dependent resources have been created. They cannot be
post-install or post-upgrade because ArgoCD will only run them once
services are healthy. But the services depend on the jobs to run to be
healthy. So have them run in the same phase as the deployment.