feat: document Dex authentication connectors
Added a new doc to explain how to use the Dex authentication connectors
to hook to Azure. Explain where the users exist in the system and how
you can add extra users for testing. Stripped down the azure config file
to just be something that needs to be appended to enable Azure
authentication.