fix: update TLS config for OpenStack
The cert-manager annotation never makes it for our public interface
override via the host_fqdn_override because that Ingress in
openstack-helm-infra does not ever have that annotation added. So the
annotation must be provided via the global Ingress annotations. The
default certificate secret for the public interface is suffixed with
public so use that as its more clear what this certificate secret is
used for.