feat: update device OBM info in Ironic
Ensure that the OBM credentials are provided to Ironic when we update
the device based on having an iDRAC or iLO interface. The Ironic API
doesn't do 'replace' operations so some extra care must be made to
detect if the values are currently set, especially for values that
cannot be retrieved like the password, to ensure we do an 'add'
operation when needed and a 'replace' when needed.