feat: create new App of Apps for ApplicationSets
As we pivot towards using ApplicationSets to control deploying
applications, we will need a new App of App to control deploying them
out utilizing annotations on the clusters. This adds the new App of App
along with adjustments to the secret generation script to create the
necessary cluster secret-type to provide those annotations. Allow the
user for now to adjust the branches or git refs used without having
ArgoCD set it back.