feat: enable extra ironic inspection hooks
Enable the LLDP, the local link connection, and the physical network
hooks for data gathering. The 'physical-network' uses IP mapping to
determine the physical network the ports are connected to.
The 'parse-lldp' option adds the LLDP info to the inventory while the
'local-link-connection' populates the baremetal ports with the LLDP
information. For the LLDP information to work, we must pass an extra
argument to the boot for it to run LLDP gathering in the first place.

https://docs.openstack.org/ironic/latest/admin/inspection.html
https://docs.openstack.org/ironic/latest/admin/inspection/managed.html
https://docs.openstack.org/ironic/latest/admin/inspection/hooks.html