neutron/undersync: change to fire & forget
We were not able to get Neutron into a state where it would wait for the
`undersync-device` workflow to complete and react to the potential
errors.

We have tried few variations:
- putting the `_move_to_network()` into a separate `greenthread` and
  periodically polling with `greenthread.sleep()`. This has made the
  workflow run in the background, but the whole Neutron instance has
  blocked
- similarly, replacing synchronous `time.sleep` with `greenthread.sleep`
  did not make any difference.
- initial Workflow trigger has been placed in the `bind_port` instead of
  `update_port_postcommit`. This also resulted in whole Neutron instance
  being blocked.

Finally we decided to simply fire the workflow and not check for the
result and rely on other mechanisms to be notified about potential
errors.

We aim to revisit this in future as it would be beneficial to interrupt
Ironic provisioning process as soon as we know that network provisioning
has failed. For that reason, the ArgoClient class has not been fully
cleaned up from the unused code.

Related docs:

https://docs.openstack.org/neutron-lib/1.9.0/reference/modules/neutron_lib.plugins.ml2.html
https://opendev.org/x/networking-arista
https://github.com/dave-tucker/odl-neutron-drivers/blob/master/odldrivers/ml2/mech_driver.py#L28
https://github.com/markmcclain/altocumulus/blob/8a82eca43ef51b9f0714baba4096a839b0b7f6e9/altocumulus/ml2/driver.py