feat: set correct ironic driver for node
If we're working with an iDRAC based interface then we need to use the
Ironic idrac driver, otherwise use the redfish based driver. this
integrates some of the behaviors of #168. The communication with
Nautobot is going to ultimately be the best approach. But the
configuration of the Ironic side is better in this code.