fix: stop using device name for interface sync
Use the device ID / UUID which is a better identifier for us to use to
find the device to sync the interfaces for. This saves us a trip to the
Nautobot DB as well. In Nautobot the device name does not have to be
unique so this avoids possible issues there. Added tests of the parser
to make sure we get the correct data. Added types so we can have better
tests in the future. Cleaned up the duplicate hostname usage in the
workflow as well.