feat: use External Secret for ArgoCD SSO auth
Used an external secret to get the ArgoCD SSO client-id/client-secret
over to ArgoCD instead of creating two secrets with the data. Broke out
the SSO piece as it's own kustomize component so that it gets applied
after the actual dependent secret is created. This solves the issue
where people had to manually restart ArgoCD because it didn't notice the
ConfigMap change and makes the install doc simpler.