feat: enable configuring nautobot from secret
The SSO secret for Nautobot contains not just the client-secret but also
the client-id and issuer which are also necessary. In the original
templating of the secret we were dropping that to just the client-secret
and then just mounting that one file. Simplify everything but just
ensuring the secret from External Secrets has all the fields. We mount
all the fields into Nautobot and then read all the fields. This keeps
our configuration in one place. Our pre-commit rules forced me to
reformat the nautobot_config.py as well.