feat: generate openstack component secrets the same
The generation of many of the component secrets was the same or should
have been the same. This moves it into a loop to preserve the same-ness
and then moves each component into a directory that can be loaded with
kustomize to avoid the awkward include need in the AppSet. Moved the
section to the top of the AppSet to match how the other AppSet's are
done.