Update Nautobot Interface not Device for provisioning mode
We used to flip a bit on the nautobot Device object for “connected_to_network”
when changing a device from/to “provisioning” and “tenant” settings.

We need this to be interface specific.  Because of Ironic's knowlege of the
interface that is being used for PXE boot, we are called with a server interface
UUID.

We update the nautobot status field of connected switch interface.

This tells undersync which mode the port should be in - provisioning or
tenant (normal).

We also make a change to the switch IDs that are passed to undersync.  This used
to represnt the set of all switches to which the server is connected.  This is
sub-optimal if there are multiple fabrics connected to the same server.   We
drop this logic and instead just pass a single switch ID, which is the switch
where PXE is happening.  Undersync will always update both switches in the VPC
pair (VLAN GROUP) in any case - passing the extra switch IDs was redundant.