Identify interface by MAC address instead of UUID
The UUID in Nautobot is not the same as the UUID in ironic/nova so use the MAC
address to find the required switch port.

In the earlier milestones we should not encounter LAGs but this is something to
watch out for in future: - it is not clear exactly what the rules are with
interface in Nautobot but when interfaces belong to a LAG.  Nautobot seems to
report that the LAG and its members all share a MAC address.  This means the
MAC address is no longer guaranteed to uniquely identify an interface.

We have also been talking about how to sync/store the various UUID values in
nautobot, so perhaps this will eventually be solved that way.