feat(openstack): use External Secrets for service account
This doesn't change any code or the actual service account that is being
used but it defines it in a way that can be consumed by external secrets
and sets the pathway for us to generate this in the future. We'd need
either an operator to action on the service account secrets to keystone
or some other way to keep them in sync.