Refactor neutron ml2 driver to make testing easier
Note that we now instantiate the ArgoClient every time it is needed, where it
used to be created at compile time and then re-used.  I think the client
initialisation is lightweight so there should be no performance penalty.  I also
think the client gathers no state between invocations so the behaviour should
stay the same.

- promote context parsing to driver level
- extract function from driver class to module top level
- inject argo client to _move_to_network function
- add docstring