Revert "neutron_understack: convert to poetry"
This reverts commits 664312d8e1828e5608671b0b2853f8c62ba0e165 and a7819cfcdc37daef433f3f30e986100f405f130c
These changes have already been implemented in:

https://github.com/rackerlabs/understack/pull/251
https://github.com/rackerlabs/understack/pull/253

Both of the PRs solve the same problem, so I don't need to solve it here
again. It was initially fixed only because I was not aware of the other
PRs. Reverting instead of dropping these commits, because they have been
referenced externally in the comments.