keystone: upgrade libapache2-mod-auth-openidc
The support for using `exec:` in Apache's `mod_auth_openidc` has been
[added][see_changelog] in version 2.4.14, so we have to install version
that is equal or newer than that. At the moment, newest version in jammy
is 2.4.11.

This commit changes the build process to install
libapache2-mod-auth-openidc out of the newer vresion of Ubuntu called
`mantic`.

Unfortunately it was not possible to install from `noble` LTS release
because of the difference in the libc6 version between those two
releases. In future, we may end up having to build this module from
source.

[see_changelog]: https://github.com/OpenIDC/mod_auth_openidc/blob/ea3af872dcdbb4634a7e541c5e8c7326dafbb090/ChangeLog#L398