fix(rabbitmq): deployment omitted an operator
Fix the deployment of rabbitmq to include both operators. Since each
operator includes a definition of the namespace, kustomize complains
about a conflict. The prior version attempted to solve that by removing
the namespace but that just caused the first operator to be omitted
entirely causing a broken deployment of rabbitmq. Lastly the
rabbitmq-cluster operator was not pinned to a version so this pins it to
the current version.