feat: disable unnecessary and broken services
OSH expects there to be two Ingress controllers in the cluster. We've
only deployed one into our setup so the second one does not exist. It is
used to be an Ingress across the internal hostname and then Services are
made pointing back to that Ingress. Those Services cannot work since
that Ingress controller does not exist. This removes those Services.
Only Keystone registered anything in the OpenStack service catalog with
that URL so this fixes that as well.