feat(understack-workflow): use clouds.yaml for auth
Provided a wrapper so that we can use clouds.yaml for authentication
from all of our workflows. Dropped the API version override from Ironic
as it conflicted with how negotiation works. Negiotation sets a minimum
and a maximum range based on server and client support. 1.82 is older
than the client and the server support so it failed. Previous
negotiation when done through the manually configured client resulted in
a range of 1.1 to 1.90, which is why we needed to pass 1.82 to get new
enough micro API. But this client results in 1.88 being used.