gitops-secrets: mute irrelevant error messages
When the understack is deployed for the first time, the script executes
a series of checks to see if particular secrets already exist. On a new
cluster, we usually don't even have the required namespace created so
check for each of the components results in a confusing "Error from
server (NotFound): namespace openstack not found". This is to be
expected, but provides no value to the user so let's silence it.