Implement password standard for BMC access
We now use a single fixed secret for the whole fleet that is used as a key to
generate a unique password for each DRAC.  The advantage here is that there is
no need to store the password for each individual host, because it can be
calculated on-the-fly.

It should be easy enough to pivot to storing passwords in vault or similar in
the future - just populate all the initial passwords with a one-liner.