fix(keystone): move groups to sso domain and have them inherit
Moved the groups to the sso domain since that's where they are really
coming from and updated them to have members of those groups inherit the
permission down to the project. This allows for the dynamic creation of
projects and people to have access to those projects. Changed the
'admin' role to the 'manager' role which is more appropriate for what
we're looking to give people.