Fetch switches from nautobot in single query and get base mac addrs
We are using nexus switches whose incorrect LLDP implementations give us the
interface mac instead of the chassis mac.  Therefore, accept either from the
switch.  Assume that nautobot always contains the "base" mac address of the
switch.

When finding the switch in nautobot that corresponds to the MAC address that was
reported to us via LLDP, first try the address verbatim, and if that is not
present in nautobot, calculate the "base" mac address (subtracting the port
number from the MAC address reported by LLDP) and try that instead.