Get Interface data from NICs that don't have the separate partition info
We were assuming that every NIC with a partition like Slot.2-1-1 also had a
redfish entry for the "base"  interface like Slot.2-1, however this is not true
on all devices.  Some only have the partition-style names.   Therefore we change
the logic: instead of including only a fixed pattern of NIC names, we ignore any
interface that looks like a partition of an existing base NIC.