fix(workflows): make ironic-node-update sensor work
The sensor didn't live in the correct namespace to be triggered by the
eventbus. The eventbus for OpenStack lives in the the openstack
namespace so the sensor needs to live there. The sensor also duplicated
the workflow calling the steps directly and now it uses the workflow
that we've created for triggering these updates. Added comments to
explain how the sensor is working and how it's triggering.