Revert "fix: fix flavor to ironic resource class conversion"
This reverts commit 5ac057b6ee2623e6542f5df3cae25be172bac718 which was
added in order to resolve the mismatch between the nova and ironic
flavors that had a punctuation in the name. The reverted commit
incorrectly changed the resource class name on the Ironic side while it
was the  Nova side that was missing the underscores. This was resolved
in the commit before this one.