openstack: adjust default PDB minAvailable
With current default setup we only spinup a single replica for each of
the components, then we apply PDB that requires this single replica to
always be available which renders the nodes the components are running
on unmaintainable.

This commit lowers the number of required available replicas to 0 with
the intention that users will override this with a more appropriate
values when scaling up the number of replicas in the deploy
configuration.