helm-charts: switch release to OCI registry
We cannot use the traditional HTTP repository hosted on the Github
Pages, because the understack repo usees github actions to publish the
website through mkdocs. The hard requirement for the
`actions/deploy-pages` action is that the artifcat that is upload is
created inside the same workflow. Currently the artifact is produced by
other mkdocs workflow so using traditional Helm Chart Releaser would
overwrite the page. We do not want that.

The potential workaround would be to:
- have the mkdocs push to a temporary branch
- have the mkdocs workflow stop pushing to Github Pages
- have the helm-chart-releaser not publish `index.yaml` to the
  `gh-pages`, but to the temporary branch instead
- additional workflow would need to be setup to react on pushes to the
  temporary branch. That workflow would need to merge outputs of mkdocs
  and index.yaml from helm chart releaser, then upload and publish to GH
  pages.

This workaround was never implemented or tested, because going down the
route of publishing the Helm chart to OCI index sounds easier as the
OCI method does not need the index.yaml.