Create associations in Nautobot to represent SVIs for external Subnets
We want to make an SVI on the leaf for every VLAN which has an
"external" subnet.

This is done via a change to the meaning of the Nautotbot data

We currently represent SVIs explicitly in Nautobot:
- we have an IpAddress object with the SVI address (.1)
- for each leaf Device, we have an Interface named "Vlan{vlan_id}"
- we associate the IpAddress with each Interface
- Interface.role is used to select the SVI configuration template

We can simplify this with a simple direct relationship between VNI and Prefix.

Using the existing Prefix->VNI mapping is a natural way to model the SVI.

- For VLANs which have a prefix, via Vlan->UCVNI->Prefix
- SVI is implicitly created wherever we create the VLAN
- bake in assumption that the SVI has .1 IP Address

This has advantages:
- data is better normalized
- fewer API or DB operations to create/delete an SVI
- keeps the "-1" and "-2" switches firmly in sync (important in a VPC pair)
- when the vlan is deleted the "SVI" will go away automatically
- easy to automatically create the SVI wherever the VLAN is used

Use VNI.role field (which has not previously been used for anything) to
specify the SVI template.

When the subnet is created in openstack

We already add the prefix to nautobot today.

However, if the subnet's "router:external" attribute is set to true, we must
perform the additional step of associating the subnet with the VNI.

When the subnet is deleted, the relationship will implicitly
cascade-delete.

We don't handle updates - it is currently assumed that no updates will
occur except for create and delete.

Note that we will not update the switches on subnet create or delete.
It is assumed that the subnet will be created before any ports are
associated with that network.

It is further assumed that is a subnet is deleted, the network will be
deleted shortly thereafter and this will perform the required cleanup.