feat: add ansible execution environment for playbooks
This creates an ansible based container that has ansible-runner as the
executor inside which can then be loaded with playbooks and roles so
that we can have a consistent way to configure different parts of the
system or react to events in the system and execute a series of steps.
To provide a consistent way to configure the overall system and react to
events in the system. Some areas where this might be useful is for
OpenStack Helm has a bootstrap bash script that executes on deployment.
This is a big hardcoded blob and will be easier to be configurable via
ansible.