feat(neutron): make our VXLAN type extend upstream
Instead of a no-op VXLAN type driver that doesn't support anything or do
anything, switch to just being an extension of the upstream VXLAN type
driver. In the future this may morph into its own distinct type so that
we can support physical network fields on this type but currently that
is not supported by Neutron. The concept of VXLAN in Neutron is centered
around the hypervisor/instance use case which does not have multiple
VRFs per hypervisor so there is no ability to separate out the VXLAN
networks. But in the physical VXLAN fabrics where different servers can
be bound to different fabrics we do have that separation.