neutron: prohibit redirect responses in Nautobot client
If the user accidentally sends the API request to the Nautobot endpoint
without trailing slash, the Nautobot will gladly accept the request and
respond with 301 redirect without processing it.
Unfortunately, the python requests library retries the HTTP request by
default without including specified payload.
So instead we simply don't allow redirects and expect the client code to
deal with a non-200 response.