fix(neutron): ensure nautobot config is passed to neutron
The neutron token isn't defined here but has been used from the
nautobot-token that is defined in the workflows which isn't good for
separation. Since 0d435bd3c7c9 the loading of the token from the file no
longer works so we must load it via a config snippet. This defines a new
secret that creates a config snippet with the token. The same change
also dropped the default URL for the Nautobot service within the cluster
so this defines that value as well.