feat: use mariadb operator for OpenStack db secrets
Generate and inject the database connection info via a secret generated
by the MariaDB operator to create a connection string that matches what
oslo.config reads.

Unfortunately OpenStack Helm doesn't always respect the mount for the
db_sync job so this won't work until that's fixed upstream and we bump
to releases with those fixes for all the services.