feat: remove OpenStack DB secrets
Since we're now using the MariaDB operator to generate the DB
connection, we don't need to have this generated by OpenStack Helm. So
that means we don't need to inject the passwords in via plain text since
OpenStack Helm does not support passing secrets by reference. Instead we
are using the db connection snippet of the prior commit.

Keystone and Placement currently do not support mounting the DB
connection snippet for the db_sync job currently so we must wait until
that is fixed upstream and we bump to a new enough version.

Placement fix: https://review.opendev.org/c/openstack/openstack-helm/+/942131