Only create SVIs for prefixes of service_type "network:understack_svi"
Previously, an SVI was created by the understack ml2 mech driver for
every subnet created with a true value for the boolean router:external
attribute.

This makes it more selective, with SVIs only being created when
service_type is equal to network:understack_svi.

The intention here is to allow for future types of "external" network
that do not require these SVIs because they have other types of router.