Don't set HTTP device (PXE) interface
We don't use this - we are leaving it at the default "disabled" mode.
On some iDRAC firmware, trying to set this while it is disabled results
in an error, which aborts the whole enroll process.

So, removing this setting will get rid of the error.   If we start using
it, and therefore "enable" the HttpDev1, this setting can be re-instated.