Fix bind_port ml2 driver interaction
We don't want to "bind" the VXLAN segments that are passed to
bind_port().  We want to make a dynamic VLAN-type segment and bind that
instead.

We are no longer using any VLAN-type networks, so remove the code that
deals with those (it was handling the provisioning network which was
still confiured as VLAN type for legacy reasons, but that is now moving
to VXLAN-type network, same as all the others.)