feat(ansible): create playbook and roles for keystone bootstrapping
The current keystone bootstrap is a bash script which makes it hard to
ensure that everything has been setup. By making it Ansible we can
ensure that things are consistent. Unfortunately the openstack's role
module does not support cross-domain or inherited like the CLI does so
this is not fully complete.