feat(ansible): create playbook and roles for keystone bootstrapping
The current keystone bootstrap is a bash script which makes it hard to
ensure that everything has been setup. By making it Ansible we can
ensure that things are consistent. Unfortunately the openstack's role
module does not support cross-domain or inherited like the CLI does so
this hacks around it with ansible.builtin.command that then needs some
hoops to avoid lint errors. Bump the keystone chart version which fixes
volume mounts so that our inventory can be mounted into the bootstrap
container.