feat(argocd): define 3 appprojects for components
This defines three different application projects for the different
component types we've got. The infra pieces are ultimately a bit
specific to the environment you're on. While cert-manager is living
there it really belongs with the operators. Operators are possible to be
cluster wide but then isn't Argo Workflows an operator? But we configure
it just to pay attention to a specific namespace so does that mean its
not an operator. Then lastly the generic understack which has the actual
applications.