feat(argocd): add new AppSet for OpenStack services
This new AppSet for the OpenStack services is a no-op with the existing
AppSet to allow for a transition over without deleting resources. This
AppSet follows the split of the regional and global deployments. It
further supports using the same AppSet definition for production
environments and development environments by using kustomize to adjust
how clusters are defined.