dexop: use dedicated instance of Dex for testing
Up till now all tests required the Dex server to be already running on
the developer's machine. It was also assumed that the server will be
preconfigured for the gRPC communication. This required the appropriate
certificates to be generated, configured in Dex and copied over to the
project directory.

This commit removes that requirement altogether by automatically
starting "dex serve" with appropriate configuration and shutting it down
when the tests complete.