chore: rename the CLI binary to understack
The intent is for this CLI tool to become a generic interface for
managing the deployment repo as well as other helper operations. The
current name of deploy-cli doesn't encompass the fact that it is part of
this project and is to specific to deployment so rename it to understack
and we will go from there.