ci: pin argo workflows binary to not get broken by release process
We've just pulled the latest argo workflows binary which has been
working but their most recent release attempt ran up against Docker rate
limits so it failed. The result is our CI is broken because the release
binary is not available. To avoid this from happening in the future, pin
the version but allow renovate to update it for us.