fix(ironic): ensure that UEFI boot works by having snponly.efi
Since we are still using iPXE with our UEFI boot, we need to have
snponly.efi available to our servers to fetch. Utilize the ironic.conf
config that causes the conductor to copy the file from the host OS,
which is located at /usr/lib/ipxe/snponly.efi in our containers, to the
path where it will serve it up from. fixes PUC-649.