feat(deploy): enable components for app deploy
The intention of the deploy repo is to provide overrides of the
defaults. The proper way to do this via kustomize is to use components
which ArgoCD gives use a better way to use like Helm values files. This
enables our deploy repo to use these components and documents how this
should appear in the deploy repo.

This depends on argoproj/argo-cd#23873