invoke undersync after the subport is commited
We have addeed subports_added_post handler that is triggered after the
validation of data is completed and the record is committed to the
neutron's database to make sure that all of the segments can be looked
up by undersync. Now the original trigger of undersync is no longer
needed.