Sync all openstack projects to nautobot, not just the user ones
This --only-domain parameter prevents projects from being created or
updated in nautobot unless they are partt of the "detault" domain.

This means the infra project is not created in nautobot, which in turn
means that out "infra" networks can't be created in nautobot.

We can't just create a user-style project to contain those networks
because rbac restrictions prevent it.