create localnet port directly in OVN
Prior to this commit we have leveraged the OVN ML2's plugin helper
method of create_provnet_port which correctly creates the port, but
unfortunately that port gets removed every time the Neutron restarts.

This is because the Neutron OVN plugin performs a database sync and when
it iterates over the LSPs that are prefixed with `provnet-` it checks if
there is a corresponding Neutron Port that is attached to the provider
network. In our case, such port does not exist so the OVN localnet port
gets removed breaking the connectivity.

Instead of relying on the helper, we execute very similar code but
prefix our LSP with `uplink-` to make it excluded from the Neutron/OVN
sync cleanup.