feat(argocd): workaround the inability to have authed components
So want to use kustomize components for overriding kustomize things.
Components work great and allow for URL based data fetching. However
within ArgoCD authenticated URLs to components do not seem to work. That
appears to be because kustomize wants URI style SSH paths while ArgoCD's
repo secrets use the SSH style.

Likely the best thing here would be to support cross-repo refs like the
helm code does but this does not appear to be supported in upstream
ArgoCD at this time.

In an effort to land the updates to the deployment process this hack was
made as a workaround. It should be retired and go away when components
are properly handled.