fix: keystone SSO group mappings were too broad
The group mappings to permission roles inside of keystone were too broad
and too much admin was applied. This scales it back to more of the
permissions that are necessary. This also changes it so that a variable
is used and can be defined to override these defaults in different
deployment scenarios.