feat(workflows): add generic script to process OSLO events
This script handles OSLO events / notifications one at a time from a
file or from stdin and dispatches it to handler functions contained
within the repo. The goal of this is to be something generic that can be
used by all of our event handling code. Connect up the Ironic port
create, update and delete events.