fix: update job hook settings for ArgoCD for OpenStack Helm
The references used here are:
https://argo-cd.readthedocs.io/en/stable/user-guide/helm/#helm-hooks
https://argo-cd.readthedocs.io/en/stable/user-guide/sync-options/#force-sync

There are certain jobs that need to run before the API deployments and
other pods from an OpenStack Helm project can start up successfully but
OpenStack Helm puts all jobs in the post sync phase which doesn't work
in ArgoCD. It works for stock OpenStack Helm because this just orders
how helm applies the manifests. They tolerate the fact that things fail
and restart a bit before coming up. But with ArgoCD it will only apply
post sync resources once the sync resources are up successfully
resulting in things never coming up. To address this we need to move
several jobs to the sync phase to complete with the deployments.

Further the jobs that OpenStack Helm creates are expected to stay around
so that it's init container can check if they exist before starting up.
Many fields in Kubernetes Job's are immutable therefore we need to
ensure that we replace the job when things change. We also don't want
ArgoCD to perform any diff (where it will error) before doing that
replacement.

ref:PUC-1082